The batch of changes for 3/2
These changes reflect the work up until 3/2. The classes have been updated to include the information required to create behavior parallel to that of the original PFEdit software. 

GUI development has been severely impeded by the inability to find example code that parallels some of the functionality we desire, such as clickable shapes, and the ability to differentiate single clicks from double clicks.